+++ fdb-sql-layer-core/src/test/java/com/foundationdb/ais/model/AISTest.java
@@ -25,7 +25,7 @@ import static org.junit.Assert.assertTrue;
 import java.util.Iterator;
 import java.util.List;
 
-import com.foundationdb.server.rowdata.SchemaFactory;
+import com.foundationdb.server.SchemaFactory;
 import org.junit.Test;
 
 public class AISTest
