+++ fdb-sql-layer-core/src/test/java/com/foundationdb/qp/row/ValuesHKeyTest.java
@@ -27,7 +27,7 @@ import com.foundationdb.ais.model.Index;
 import com.foundationdb.ais.model.Table;
 import com.foundationdb.ais.model.TestAISBuilder;
 import com.foundationdb.qp.rowtype.Schema;
-import com.foundationdb.server.rowdata.SchemaFactory;
+import com.foundationdb.server.SchemaFactory;
 import com.foundationdb.server.types.service.TypesRegistryService;
 import com.foundationdb.server.types.service.TypesRegistryServiceImpl;
 import com.persistit.Key;
