+++ fdb-sql-layer-core/src/test/java/com/foundationdb/server/service/externaldata/CsvRowReaderTest.java
@@ -21,7 +21,7 @@ import com.foundationdb.ais.model.AkibanInformationSchema;
 import com.foundationdb.ais.model.Column;
 import com.foundationdb.ais.model.Table;
 import com.foundationdb.qp.row.Row;
-import com.foundationdb.server.rowdata.SchemaFactory;
+import com.foundationdb.server.SchemaFactory;
 import com.foundationdb.server.types.mcompat.mtypes.MTypesTranslator;
 import com.foundationdb.server.types.value.ValueSources;
 import com.foundationdb.util.Strings;
