+++ fdb-sql-layer-core/src/test/java/com/foundationdb/server/test/ApiTestBase.java
@@ -50,6 +50,7 @@ import com.foundationdb.qp.rowtype.IndexRowType;
 import com.foundationdb.qp.rowtype.RowType;
 import com.foundationdb.qp.util.SchemaCache;
 import com.foundationdb.server.service.blob.LobService;
+import com.foundationdb.server.SchemaFactory;
 import com.foundationdb.server.store.FDBHolder;
 import com.foundationdb.server.store.FDBStore;
 import com.foundationdb.ais.AISCloner;
@@ -72,7 +73,6 @@ import com.foundationdb.server.types.value.ValueSource;
 import com.foundationdb.server.types.value.ValueSources;
 import com.foundationdb.sql.LayerInfoInterface;
 import com.foundationdb.server.AkServerUtil;
-import com.foundationdb.server.rowdata.SchemaFactory;
 import com.foundationdb.server.service.ServiceManagerImpl;
 import com.foundationdb.server.service.config.ConfigurationService;
 import com.foundationdb.server.service.config.TestConfigService;
