+++ fdb-sql-layer-core/src/test/java/com/foundationdb/server/types/ConversionHelperBigDecimalTest.java
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package com.foundationdb.server.rowdata;
+package com.foundationdb.server.types;
 
 import org.junit.Test;
